♻️ refactor : 도메인 폴더 이동에 따른 의존성 변경
